modify extensions rule to apply for all exports
